--- src/NewApiBundle/Entity/Commodity.php
@@ -1,10 +1,0 @@
-<?php
-
-
-namespace NewApiBundle\Entity;
-
-
-class Commodity extends \DistributionBundle\Entity\Commodity
-{
-
-}
