--- src/NewApiBundle/Mapper/BeneficiaryMapper.php
@@ -1,10 +1,0 @@
-<?php
-
-
-namespace NewApiBundle\Mapper;
-
-
-class BeneficiaryMapper extends \BeneficiaryBundle\Mapper\BeneficiaryMapper
-{
-
-}
