+++ src/NewApiBundle/Resources/config/services.yml
@@ -4,17 +4,6 @@ services:
     autoconfigure: true
     public: false
 
-  NewApiBundle\Model\HomeService:
-  NewApiBundle\Model\AssistanceService:
-
   NewApiBundle\Mapper\AssistanceMapper:
-  NewApiBundle\Mapper\BeneficiaryMapper:
-  NewApiBundle\Mapper\CommodityMapper:
 
-  NewApiBundle\Repository\AssistanceRepository:
-    factory: ['@doctrine.orm.default_entity_manager', getRepository]
-    arguments:
-      - NewApiBundle\Entity\Assistance
-
-  NewApiBundle\Controller\HomeController:
   NewApiBundle\Controller\AssistanceController:
